Separated more files into API and implementation to enable use from LearnLib.
