+++ api/src/main/java/net/automatalib/automaton/time/mmlt/AbstractSymbolCombiner.java
