+++ api/src/main/java/net/automatalib/automaton/time/mmlt/semantics/TimeoutPair.java
